move 'navigate' code to separate directory
